добавил докер и докеркомпоз файлы и поправил ключ bandwidth

+++ my_app/operations/models.py
@@ -9,4 +9,4 @@ class QueryInfo(Base):
     address = Column(String, index=True)
     balance = Column(Float)
     energy = Column(BigInteger)
-    bandwidth = Column(Integer)
+    bandwidth = Column(BigInteger)
